requirements.txt
agate==1.13.0
agate-dbf==0.2.3
agate-excel==0.4.1
agate-sql==0.7.2
asgiref==3.8.1
babel==2.17.0
csvkit==2.0.1
dbfread==2.0.7
Django==5.1.6
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
et_xmlfile==2.0.0
greenlet==3.1.1
isodate==0.7.2
leather==0.4.0
numpy==2.2.3
olefile==0.47
openpyxl==3.1.5
pandas==2.2.3
parsedatetime==2.6
phonenumbers==8.13.55
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyaes==1.6.1
pyasn1==0.6.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.38
sqlparse==0.5.3
Telethon==1.39.0
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
xlrd==2.0.1
